Eric Sunshine eliminated some additional compilation warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22540 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/lighter2/kdtree.h
@@ -44,7 +44,7 @@ namespace lighter
     RadPrimitivePtrArray radPrimitives;
 
     KDTreeNode ()
-      : splitDimension (CS_AXIS_NONE), splitLocation (0.0f),
+      : splitDimension ((uint)CS_AXIS_NONE), splitLocation (0.0f),
       leftChild (0), rightChild (0)
     {}
